refactor: move task-related events into events/tasks. Give eventType and config individully
