--- src/libs/events/EventType.js
@@ -1,24 +1,0 @@
-// 分享薪資工時 CreateSalaryWorkTime 任務完成 handler
-const createSalaryWorkTime = "CreateSalaryWorkTime";
-
-// 分享面試經驗 CreateInterviewExperience 任務完成 handler
-const createInterviewExperience = "CreateInterviewExperience";
-
-// 分享工作心得 CreateWorkExperience 任務完成 handler
-const createWorkExperience = "CreateWorkExperience";
-
-// 分享網站到 shareWebsite 任務完成 handler
-const shareWebsite = "ShareWebsite";
-
-const unlockExperience = "UnlockExperience";
-
-const unlockSalaryWorkTime = "UnlockSalaryWorkTime";
-
-module.exports = {
-    createSalaryWorkTime,
-    createInterviewExperience,
-    createWorkExperience,
-    shareWebsite,
-    unlockExperience,
-    unlockSalaryWorkTime,
-};
