+++ src/libs/events/rewards/EventType.js
@@ -1,0 +1,8 @@
+const unlockExperience = "UnlockExperience";
+
+const unlockSalaryWorkTime = "UnlockSalaryWorkTime";
+
+module.exports = {
+    unlockExperience,
+    unlockSalaryWorkTime,
+};
