+++ src/libs/events/tasks/EventType.js
@@ -1,0 +1,18 @@
+// 分享薪資工時 CreateSalaryWorkTime
+const createSalaryWorkTime = "CreateSalaryWorkTime";
+
+// 分享面試經驗 CreateInterviewExperience
+const createInterviewExperience = "CreateInterviewExperience";
+
+// 分享工作心得 CreateWorkExperience
+const createWorkExperience = "CreateWorkExperience";
+
+// 分享網站到 shareWebsite
+const shareWebsite = "ShareWebsite";
+
+module.exports = {
+    createSalaryWorkTime,
+    createInterviewExperience,
+    createWorkExperience,
+    shareWebsite,
+};
