+++ src/routes/interview_experiences/index.test.js
@@ -10,7 +10,7 @@ const {
     COMPLETED,
 } = require("../../models/schemas/userPointEvent");
 const { createInterviewExperience } = require("../../libs/events/EventType");
-const taskConfig = require("../../libs/events/task_config");
+const taskConfig = require("../../libs/events/tasks/config");
 
 const app = require("../../app");
 chai.use(chaiAsPromised);
