+++ src/schema/experience.test.js
@@ -12,7 +12,7 @@ const {
     createInterviewExperience,
     createWorkExperience,
 } = require("../libs/events/EventType");
-const taskConfig = require("../libs/events/task_config");
+const taskConfig = require("../libs/events/tasks/config");
 
 chai.use(require("chai-subset"));
 const { assert, expect } = chai;
